Change SalaryValue from Long to BigDecimal

+++ backend/src/main/java/com/theladders/api/dto/JobSalaryRequest.java
@@ -1,7 +1,9 @@
 package com.theladders.api.dto;
 
+import java.math.BigDecimal;
+
 public record JobSalaryRequest(
-        Long value,
+        BigDecimal value,
         String currency,
         String unit
 ) {
